+++ packages/@dopt/ai-assistant/definition/src/api/completions.ts
@@ -16,9 +16,7 @@ export type UserIdentifierParams = Static<typeof UserIdentifierParams>;
 
 export const AssistantQueryParams = Type.Object({
   ...UserIdentifierParams.properties,
-  model: Type.Optional(
-    Type.Union([Type.Literal('gemini'), Type.Literal('gpt')])
-  ),
+  model: Type.Optional(Type.String()),
 });
 export type AssistantQueryParams = Static<typeof AssistantQueryParams>;